LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update-module-request-mode.https-expected.txt


FAIL headers of a main module script assert_equals: expected (string) "same-origin" but got (undefined) undefined
